requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # required for file upload handling

# ── ML / NLP ──────────────────────────────────────────────────────────────────
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4

# ── PDF Parsing ───────────────────────────────────────────────────────────────
PyMuPDF==1.24.3                # primary PDF parser (fast)
pdfminer.six==20221105         # fallback PDF parser

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.7.1

# ── Optional: Semantic Search (uncomment to enable) ───────────────────────────
# sentence-transformers==2.7.0
# faiss-cpu==1.8.0
